hcitool backend: poll on hcidump (not hcitool)
The _start_scan_hcitool() loop should poll on the hcidump subprocess, not
the hcitool subprocess.  The hcitool command doesn't work on BlueZ 5 so
it immediately fails.  It may still be needed on Bluez 4.  The main
loop should continue as long as hcidump is still running and ignore the
status of hcitool.